+++ src/django_upgrade/fixers/versioned_test_skip_decorators.py
@@ -31,7 +31,7 @@ from tokenize_rt import Offset
 
 from django_upgrade.ast import ast_start_offset, is_passing_comparison
 from django_upgrade.data import Fixer, State, TokenFunc
-from django_upgrade.tokens import erase_decorator
+from django_upgrade.tokens import erase_decorated_def, erase_decorator
 
 fixer = Fixer(
     __name__,
@@ -108,10 +108,25 @@ def _handle_decorator(
                 (pass_fail := is_passing_comparison(decorator.args[0], state))
                 is not None
             )
-            and (
-                (ident == ("unittest", "skipIf") and pass_fail == "fail")
-                or (ident == ("unittest", "skipUnless") and pass_fail == "pass")
-                or (ident == ("pytest", "mark.skipif") and pass_fail == "fail")
-            )
         ):
-            yield ast_start_offset(decorator), partial(erase_decorator, node=decorator)
+            # If condition is always true, so the decorated def is always skipped and can be removed.
+            always_skipped = (
+                (ident == ("unittest", "skipIf") and pass_fail == "pass")
+                or (ident == ("unittest", "skipUnless") and pass_fail == "fail")
+                or (ident == ("pytest", "mark.skipif") and pass_fail == "pass")
+            )
+            if always_skipped:
+                yield (
+                    ast_start_offset(node.decorator_list[0]),
+                    partial(
+                        erase_decorated_def,
+                        first_decorator=node.decorator_list[0],
+                        node=node,
+                    ),
+                )
+                break
+            else:
+                yield (
+                    ast_start_offset(decorator),
+                    partial(erase_decorator, node=decorator),
+                )